requirements.txt
alembic==1.5.7
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.7.0
Flask-MySQL==1.5.2
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
mysqlclient @ file:///C:/Users/nsest/Downloads/mysqlclient-1.4.6-cp38-cp38-win32.whl
PyMySQL==1.0.2
python-dateutil==2.8.1
python-editor==1.0.4
requests==2.25.1
six==1.15.0
SQLAlchemy==1.4.1
urllib3==1.26.4
Werkzeug==1.0.1
